pyproject.toml
[tool.poetry]
name = "family_account"
version = "0.1.0"
description = ""
authors = ["Skrillite <nazarov.k.k@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
sanic = "^21.6.0"
SQLAlchemy = "^1.4.20"
python-dotenv = "^0.18.0"
pydantic = "^1.8.2"
asyncpg = "^0.23.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.6b0"
alembic = "^1.6.5"
pytest-aiohttp = "^0.3.0"
pytest-asyncio = "^0.15.1"
pytest-coverage = "^0.0"
pytest-dotenv = "^0.5.2"
pytest-pythonpath = "^0.7.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
